refactor: ログイン, 外部認証等の処理のリファクタリング
